+++ lisp/eab-ui.el
@@ -162,6 +162,8 @@
 (setq browse-url-browser-function (quote browse-url-firefox))
 (setq browse-url-firefox-program "/usr/local/bin/browser-remote")
 
+(setq eww-search-prefix "https://duckduckgo.com/html/?q=")
+
 (defun google ()
   "Google the selected region if any, display a query prompt otherwise."
   (interactive)